Re-ordering of Previous File/Next File links
The links now follow the order and filtering parameters set by the user
when visiting the file index. The previous/next links to files of the
same type has been removed, as that functionality is now convered by
filtering the index by type and then using the regular previous/next
links. Also, as to make the functionality more intuitive, the links now
bear the name of the file they point to.