Issue #81 - Uselessness of switch_panel and inline_edit_form
Both switch_panel and inline_edit_form have been removed, with their
associated elements (a partial and a JS event handler), respectively.

The switch_panel functionality was removed as part of what seems to be a
refactor of the interface, see commit a526dea. There is no reference
to the partial or method anywhere else.

Searching through the commit history for inline_edit_form reveals the method
never had any callers and is part of the inline_edit_* JS-based rich edition
functionality. It was introduced in commit d9df961 as a complement of
inline_edit_field, which is still used in
BrainPortal/app/views/userfiles/show.html.erb, line 51