Issue #15 - Framework for help documentation
Help documentation controller, model and views. The documentation is
stored in Markdown format under Rails.root/public/doc and HTML
conversion is handled by the Showdown JS library while edition
(reserved to the core admin) uses a JS-based UI based on CodeMirror.

To add help buttons using this framework, use the help_button helper
defined in DocumentationHelper with a key indicating which help
document is to be shown.