Issue #77 (partial) - Download File button leads to an error page
This problem occurs when the data provider is marked online but is not
actually alive (it doesn't respond). As checking if the data provider is
alive can take a while, this changeset just makes sure a (somewhat) proper
error is shown to the user if the data provider is dead.