Issue #92 - Killing a subprocess no longer works
As described in the issue comments, Ruby 2.0+ raises a ThreadError when
trying to use a lock in a trap context. This changeset works around that
problem by collecting log lines (as the logger used, log4r, uses a lock
to ensure thread safety) and flushing them out to the logger at the end
of the trap callback by using a separate thread.