Issue 6903 [redmine] - Consistency checks for data providers
The checks are implemented using 2 new data provider API methods;
provider_report and provider_repair. provider_report returns a list of
issues (hashes) representing consistency problems in the provider, while
provider_repair is given an issue and tries to automatically correct it.

Following convention, the subclass implementation of provider_report and
provider_repair are in impl_provider_report and impl_provider_repair.

A sensible default of impl_provider_repair is provided; it tries to
solve the issue based on the issue's :action key.