Fix for issue #120 - Cannot sort exceptions by date, revision...
Caused by the session code's messiness; trying to default to ascensing
order by using an empty string causes it to be removed later by a
delete_if on blank?s. This makes the controller (and view) think that
the sorting order was never set in the first place. The exception and
feedback controllers then legimately want to sort in descending order
by default, triggering the bug when a request is made to sort the other
way.