Fix for selection state on persistently selected files
This changeset fixes a bug where persistently selected file rows would
not be selected in the main userfiles table. This bug was introduced
in the port of the table to new dynamic tables, which at the time did
not support specifying a row's selection state.

This changeset also reworks the userfiles_display partial's style to
match the one used for most index tables.