Disabled version switching, minor fixes
The currently implemented way of specializing the 'version switcher'
instance to the version-specific class (bulk-importing methods) crashes
Ruby when trying to garbage-collect the object.

At the time of writing, there is no known good way to implemented the
required functionality in Ruby without resorting to even deeper and
messier hacks. As such, the functionality has been disabled for now (its
off by default, and neither the auto-loader or Rails generator enable
it).

This changeset also adds a warning when a to-be-generated task class has
already been defined before and corrects some minor issues.