Session data controller, CbrainSession refactor to reduce coupling (!)
CbrainSession now acts as an independent model wrapping
ActiveRecord::SessionStore::Session, and no longer concerns itself with:
- Session value format; the undocumented API relying on special suffixes
  (_clear, _remove, _hash, _array, etc.) in session keys has been
  removed.
- Controller-specific keys; CbrainSession no longer directly considers
  keys named from controllers (bar a single method kept for backwards
  compatibility.
- Filtering and sorting; all methods related to sorting and filtering
  have been removed, as they should belong in their own module or class.
All methods have been re-worked to be as simple as possible, documented
and meet CBRAIN's coding style.

This changeset also adds a new controller to allow direct API access to
session data, which is intended to be used from client-side JS scripts.

Note that as this changeset removes the filtering/sorting methods from
CbrainSession, it breaks the filtering and sorting capabilities offered
by the basic_filter_helpers and filter_helper modules. This
functionality will be restored by the upcoming refactor of those
modules.