Scope-backed dynamic_table helper
The new helper, dynamic_scoped_table, is similar in behavior to
dynamic_index_table; it extends dynamic_table with native integration
for the new scope-based filtering mechanism; the options `order_map`
and `filter_map` can be used to specify which Order and Filter instances
to use for sorting and filtering links while `scope` corresponds to the
currently active scope to fetch the current sorting order and target
scope name from.

This changeset also include a few bugfixes for dynamic table's ERB
template.