Port of most controllers and views to the Scopes API
This changeset ports most controllers and views to use the new Scopes
mechanism instead of the basic filters one. The recently ported
controllers and views have a few differences stemming from the API
change:
- Total filter counts are no longer available (TODO), and available
  filter values correspond to the currently shown object list only.
- There is no more @header_scope, @filtered_scope and @filter_params;
  they have been replaced by @view_scope, which corresponds to
  @filtered_scope and @scope, which somewhat corresponds to
  @filter_params.
- The messages index has been reworked, as the old version was somewhat
  clunky and, most importantly, hard to port correctly.

This changeset also refactors some previously ported controllers and
views to use the new helpers introduced in the previous commit (extra
helpers).

Note that this changeset most likely horribly breaks a bunch of things,
as it is very wide reaching and completely refactors a key component in
many places.