Refactor of userfiles persistent selection
This changeset refactors the userfiles to use the new persistent
selection mechanism and all its new features (and quirks). It also
removes the old one including all references to the
:persistent_userfiles session information.

Note that as the persistent selection is now stored on the client, it is
no longer practical for the server (usually controllers) to clear it.

This changeset also applies some minor style fixes to the persistent
selection component.