Issue #276 - Page number out of bounds
Fixed by explicitly computing the size of the result set (instead of
using a static value) to clamp the page number. This fix naturally
obsoletes the need to reset the page to 1 when invoking a dynamic table
filter.

Note that computing the result set's size usually involves an extra
database query and thus has a non-negligible performance impact.