Missing quality control (QC) button/item
Added a menu item in the 'more' menu invoking the QC panel, behaving
similarly to the old UI's oddly placed 'Start QC' button. This menu
item is a late (and hopefully temporary) addition to the new UI as
QC functionality does not really belong there (no other suitable
location was found for the time being).