Issue #260 - Check exit status of descriptor-generated tasks
The exit status (or return code) of generated ClusterTasks is now
checked by default using a small file (.qsub.exit.<...>) and a
non-successful exit status forces the task's status to 'Failed On
Cluster'.

To disable this behavior, set the flag 'cbrain:ignore-exit-status'
as true in the task's descriptor.

This changeset also reworks some odd styling in the bourreau.rb.erb
template.