Issue #290 - Broken name search in userfiles index
Caused by improper handling of the custom filters array (:custom_filters
in the userfiles scope) when unset; the &= assignment in custom_scope
would incorrectly assign a value of 'false' to :custom_filters, breaking
the _active_filters partial as it would try to interpret this value as a
custom filter id.

Fixed by strengthening both _active_filters and custom_scope, by
allowing the former to ignore incorrect custom filter entries and
explicitly setting :custom_filters in the latter to an empty array
(if unset).