Issue #302 - Stale selection count
This changeset addresses a bug which caused the userfiles' dynamic menu
to appear even if no files were selected; the persistent selection
component's psel-count element would not be up-to-date when queried
and would mistakenly return an higher value.

Fixed by reworking persistent selection events to have an event fire
just after the component is finished updating. This event is then
handled to update the dynamic menu's status.

This changeset also fixes a few minor quirks in
_selection_database.js.erb.