Improved persistent selection requests
Persistent selection requests no longer send all bound elements; they
instead send which elements were unselected by the user. In the worst
case, the request is just as big as with bound elements, and in the
average case the request is much smaller, especially with a large number
of bound elements (userfiles tables @ 1000 per page, for example). Note
that this improvement comes with a slight slow down when selecting
elements, as the unselected elements list needs to be maintained.

This changeset also tries to fix an odd bug where .html gets appended to
requests, causing some server-side JS requests to return HTML instead of
the expected JS.