Issue #320 - Safe AR inspect method
This monkey-patch of AR's inspect method on relations avoids
accidentally instantiating tons of AR records and filling up all
available memory if the relation is too big (> 5000 records). The patch
is rather restrictive; in some cases, invoking inspect is fine even if
the relation is big.

The original bug leading to this changeset was an invalid method name in
one if the userfile views, raising a NameError which would invoke
inspect on every instance variable. One of those instance variables was
the base userfile scope, an AR relation representing every userfile
available to a given user. Trying to inspect that relation would try to
instantiate every userfile in the database (> 160000) and consume all
available memory.