Issue #156 - Processing of read-only files
A new property, :readonly_input_files, indicates if a task only needs
read access to its input files, allowing users to launch tasks on files
they only have read access to. Note that this new property is only
relevant for the portal (as it is the one to perform the check).

This changeset also contains a minor fix for task directives and updates
the descriptor-based task generator to explicitly show task properties.