Better metadata values for DP access restrictions
More test-friendly values for some access checkboxes. The actual
values in the metadat is not important, since what only matters
is if the metadata key is present or not. So instead of having
a value of "disabled" I put a more descriptive sentence about what
it means (e.g. "remote resource X cannot sync DP Y"). It also
makes inspecting the meta data more friendly with the console.