Issue #330 - Empty session with _simple_filters
Trickier _simple_filters handling to cleanly (somewhat) allow new scopes
to be created when the session is empty. The new behavior is for a scope
to be created with _simple_filters's value when none can be found
(including other defaults), falling back to default_scope_name if
_simple_filters's value looks like '1' or 'true'.

While hard to precisely define, this behavior should hopefully be
simpler to use and fix the current issue.