No sessions created for service requests.
This commit changes the behavior of the logging
code that's run after each request such that no
writing is done to the session at all. There used
to be a variable 'client_type' that was set in one
of the after_filter callback, log_user_info() to
which was also used by the afgter_filter callback
action_counter(), and just the creation of this entry
cause a session to be created. This was annoying
when being monitored externally using the service
controller.