Issue #335 - Missing Bourreau load info in tooltip
Caused by a rather old rework of the load_info action switching from
directly loading the load info in the view to passing it from the
controller, causing issues as the load_info partial is used in other
places (see issue discussion for the exact commit and more details).

Fixed by re-adding the load info logic as a default in the view. Not the
most elegant solution, but the queries in question are quite simple and
the duplication hopefully won't be too big of a code smell...

This changeset also 'fixes' a strange UI bug where an UI element related
to tool selection would show up in the load info tooltip.