Issue #18 - Refactor of DataProvidersController's register action
The action is now split by the three actions it performs; register,
unregister and delete/erase, and most work is not done in background to
avoid slowing down the request (non-API requests only). Most oddities
have been ironed out, and the error reporting process now follows more
closely the one used in other parts of CBRAIN for processing lists of
files.

Code duplication has been avoided as much as possible (see the lot of
private internal/utility methods added at the bottom of the file), but
some parts of the three actions do end up being very similar.

These changes should not have affected the API responses or HTML request
parameters required from the UI, save for the use of new actions which
employ the same parameters as the previous all-in-one action.