Issue #349 - Select-all checkbox off by a few pixels
Caused by a base stylesheet difference between Chrome and Firefox;
Chrome defaults to 'center' while Firefox defaults to 'start' (which
translates to 'left'). Fixed by manually forcing 'left' in
dynamic-table.css.

This changeset also makes the checkbox column a fixed-width column (as
much as possible), as it looks nicer when the table's columns are
stretched out.