Added AccessProfile mechanism.
This is a new model that is only useful for the administrator.
Each AccessProfile is basically

* a list of groups (a.k.a projects)
* a list of users

The purpose is to simplify user management. The admin
creates a AccessProfile, assigns it a bunch of groups,
and then when creating users or modifying them, they
can be assigned to the profile. When the profile's
list of group is changed, all users associated with the
profile will have their list of groups changed too.

Users can be assigned multiple AccessProfiles, which means
they get the UNION set of all groups from those profiles.

AccessProfiles have pretty labels the show up as colorful
boxes for the admin.