Cleaned up and restructured the DP class tree.
I've introduced three new DP classes:

* FlatDirLocalDataProvider
* FlatDirSshDataProvider
* FlatDirSmartDataProvider

They are browsable with the files all in
one level of a directory; this is similar to
the SshDataProvider but with two different
transport mechanism (local and ssh). In fact,
the FlatDirSshDataProvider is just a subclass
of SshDataProvider with nothing new in it, since
all the function was already in there.

I've improved the content of LocalDataProvider
to provide a proper base set of methods for building
local DPs. It almost (but not exactly) implements
the full set of methods for FlatDirLocalDataProvider.