Fixed the timing logic for synchronization.
It seems opening a transaction on a Userfile in
data_provider.rb was making the associated SyncStatus
object invisible to all other DB clients. Changed
the code so we only lock state using SyncStatus itself,
as it was supposed to be. There might still be very very
rare race conditions, but the use of status_transition!()
will help detect that.