Fix: same content registered twice on some DPs
On data providers using a flat directory layout, it was
sometimes possible for two users to register files
with the same names, and then get their contents mixed up.
This is prevented by new callbacks, and also the upload
operation has been modified to handle this a bit more
gracefully.