Support for revinfo in plugins.
Plugins can now provide a flat file of revision info
in their top directory level, which will be parsed
just like for the main platform, when the app is not
deployed with git.

Script have been added/modified to CBRAIN's "script" subdirectory
to help (re)generate all the flatfiles.

The interface show the revsision info for the plugins in
the "about us" page.

Two unrelated files get better doc (in app/models).