Modal for DELETE FILES example fix.
This commit contains the minimum amount of changes
necessary to make the UI interface work with the
bootstrap modals. use this as an example on how
to do the other buttons.

Note that the original dialog code had the ability to
make custom dialogs for buttons, or do make much simpler
YES/NO dialogs simply by specifying appropriate attributes
in the buttons. For bootstrap we will simply make an entire
dialog for each button, even if it's only yes/no.