Tasks are started with a wrapper bash script.
This allows us to view live the output of the scientific
script.

The rest of the API has been kept perfectly compatible.
One added behavior is that the framework, at task post processing,
will check for a particular sentence echoed back at the end of the wrapper
script to make sure the task ran properly to the end, and
otherwise will mark the task as Failed On Cluster immediately.