Travis CI improved: DB inside container.
This commit changes completely the manner
that Travis CI is integrated within CBRAIN.
The docker container that we build separately
now contains a fully seeded DB, and when running
the test we do the minimum amount of work
necessary to bring the internal CBRAIN install
up to par. We only bundle when Gemfiles have changed;
we only migrate when the migration files have changed,
we only install the plugins when their files have changed.