Fixes some tests.
The failures were related to allow() and expect() statements
that were made before the FactoryGirl construction of some
objects were performed, and these interfered with those constructions.
Assigning the constructed objects to variables BEFORE specifying
the test conditions solved that. It's something to remember, in
general.

Also, failures are limited to 50 on each Portal and Bourreau sides.