API query limits. Fixes #668, #697
Fix for #687 :

The method for_api() was added to ActiveRecord::Relation;
it checks if the relation is limited, and if not imposes a limit.
Then it invokes to_a() and then Array's own for_api().

Fix for #668 :

At the same time, moved all the extensions we had defined in
ActiveRecord::Base back into ApplicationRecord.