Updated the old s3_provider to use the new AWS SDK lib.
Renamed the s3 connection library to distinguish it from the old one.

Eliminated lib/s3_connection.rb as that is not used anymore and doesn't work.

Renamed the new S3 Data Provider to S3FlatDataProvider to be more accurate with its functionality and
distinguish it from the old one.

Added text to the new.html and description of the dataproviders just to make sure it was clear.