Better handling of IP addresses
We use Rail's request.remote_ip instead of our own.

When users roam, the IP address and host name will be updated.
In the past, we always showed the IP address that the user
initialy used when logging in.

Also, requests made with API tokens are checked for their
IP addresses, and if they change, the session is de-activated.