Bourreau workers propagate stop to subworkers
There was a bug that prevented the stop signal from
propagating: basically if logger was invoked from
the signal handler, it crashed and the rest of the signal
handler code was ignored.

The main worker class already had code to mitigate this,
with @trap_log a variable holding log messages that
are sent at the end. So the subworker use that too now.