Added resource usage tracking.
We have a now a generic structure for recording resource usage changes.
It is an ActiveRecord of class ResourceUsage, with several subclasses:

  ResourceUsage (abstract)
  |
  +- SpaceResourceUsage (abstract)
  |   |
  |   +- SpaceResourceUsageForCbrainTask
  |   |
  |   +- SpaceResourceUsageForUserfile
  |
  +- TimeResourceUsage (abstract)
      |
      +- CputimeResourceUsageForCbrainTask
      |
      +- WalltimeResourceUsageForCbrainTask

* Whenever a file's size is updated, one SpaceResourceUsageForUserfile
  is created.

* Whenever a task transitions to 'Data Ready', one
  WalltimeResourceUsageForCbrainTask will be created and one
  CputimeResourceUsageForCbrainTask will be created.

* Whenever a task reach a final state, one SpaceResourceUsageForCbrainTask
  record will be created.

All records contain optional associations to the userfiles, tasks,
tools, tool_configs etc. For all of these assocations, because we
want to keep the usage information even after the associated abject
is deleted, we also keep a full copy of some attributes. E.g. for
a task, we keep its type, its status, its tool name, etc.

For the moment, the migration doesn't include the creation of any indexes.