Bourreaux report their expected number of workers
This will simplify nagios checkers, they'll be able
to verify that the number of active workers match the
expected number. Right now the number was passed in
argument to the nagios script, which meant we needed
to change the nagios config every time we changed the
number of workers on a cluster.