Better navigation between CBRAIN <-> NeuroHub
Also:

* we record if the user submitted a new account request
using the CBRAIN form or the NeuroHub form. The info
is shown to the admin, that's all.

* we record if the user connected using the CBRAIN login
page or the NeuroHub login page; this info is not used right now
but could be turned into a "return home" link, as generated
by the new helper code (see login_dashboard_path(), not
used right now)