Portal boot and runtime info enhancements
When the portal boot in console mode, the admin is now
allowed to use a CBRAIN main SSH key that has a passphrase.
The boot process will stop and ask for it. If not provided,
it continues with a warning. Once a key is added once, there
won't be a need to supply the passphrase again as long as the
SSH agent is active.

Tracking runtime information in tasks now includes
the output of 'numactl -s'.