Removed some SSH config options from Master
The default set of options when setting up masters
included stuff for disabling keyboard auth.
But this was now starting to interfere with more recent default
installation of SSHd which refuse to connect with
SSH keys even when those options were specified,
eeven if we're never even attempting authentication
by keyboard.