Data Tracking Framework.
This allows an administrator to mark a WorkProject
with a 'data tracking' flag. When enabled, the CBRAIN
codebase will count, monthly, the files of those
projects as used in four contexts:

1) downloads using the interface
2) views in the browser
3) copies made to other data providers
4) files used to setup a CBRAIN task

For each context, a value of +1 is added to the usage count
and and value of +num_files is added to a num_file counter
(so basically we can see the number of files used in file
collections).

The standard report maker of CBRAIN has been extended to add
four new report types for these values.