First pass at upgrading to UNIX sockets
This commit removes the old TCP scheme
between portal and bourreaux, and replace it with
a similar one based on UNIX domain sockets. These
are still forwarded by a SSH master.

I ripped out the capability to connect to a second-level
submit host, however. The remote host we connect to, where
the Bourreau runs, is assumed to have the ability to submit
jobs.