Turn off cbrain:no-run-id-for-outputs on flat DPs
The Boutiques option 'cbrain:no-run-id-for-outputs', a custom
option in the descriptor, normally allows the integrator to
say that outputs created will not contain a "-1234" task ID
suffix. This can cause a task to crush its own input (or some
irrelevant files) at post processing. The option is now only
implemented when the destination data provider is MultiLevel,
presumably because in that case the output will go to a distinct
repository by using BoutiquesForcedOutputBrowsePath.