Improvements to BAC framework
The Worker now logs when a BAC is finished.

An API entrypoint as been added so that a user can get
information about a BAC (JSON only).

A new helper method userfile.unregister has been added which
destroys the userfile but leaves the file's content on the
DP side. This is allowed only when the DP is browsable, of
course.