Added new Nagios checker
New rake task to ceheck the amount of disk space used on each
bourreau. Invoke like this:

  rake cbrain:nagios:checkers:workdirs <clustername>:<warn>:<crit>

where clustername is either the ID or name of a bourreau, warn
is an integer representing the number of GB to trigger a warning
and crit the number of GB to trigger a critical. E.g.

  rake cbrain:nagios:checkers:workdirs Beluga:2000:5000

will be in warning after 2TB is used, and will be critical after 5TB.