No longer use a lockfile for the AgentLocker subprocess
We used to have a piece of code that created a lockfile to prevent
multiple instances of the web server from each starting a copy of
the AgentLocker subprocess, but that was when Rails used multiple
servers. With puma, we start only a single initial server which
then forks after initialization, so we no longer need that code.