CleanCache performed by BACs now.
The old CleanCache operation was performed by sending
a message and having the remote resource spawn(). Now
the remote resource schedule BackgroundActivity objs instead.

A better improvement, eventually, would be for the BACs to
be created on the portal side without even having a message sent.
So, something to think about.