Changed BoutiqueSupport class declarations
The old code used meta programming to create the
RestrictedHash subclasses that hold the boutiques
structures. This new code doesn't change anything
in the functionality, but does it differently,
in a more classic manner.

This change allow us to add proper documention
in rdoc format on all the internal methods (it was
really difficult otherwise). Doc will be provided later.