No longer tries streaming 'within' SingleFiles
The stream method sends a SingleFile has is; it
sends any file inside a FileCollection using file_path
as a relative path. The previous code ignored
file_path for SingleFile and always streamed the SingleFile
itself. Now if file_path is provided for a SingleFile and
it doesn't match the name of the SingleFile, :not_found is
returned.