Initial implementation of a CPU quota framework (#1498)
A big part of the existing DiskQuota system was re-used.
The controller and models were renamed and extended.

Currently, this framework is operational except for the part where
quotas are actually enforced. So when deployed, we can create and
query quotas, but no user will be prevented from running tasks if their
quotas are exceeded.

I will add later the enforcement, because I want to see how the quota
objects behave in the real system.