ToolConfigurator, fixes #1422 (#1508)
This PR implements a new admin-only feature. A CBRAIN task can be launched
to build a container for a ToolConfig. The task launch form is also used as
a convenience "copy attributes" helper so that the admin can copy special
configuration options from an existing ToolConfig (e.g when configuring
a tool 1.0.5 when 1.0.4 already exists, we can copy the squashfs configurations,
slurm parameters, project ID, etc etc).