Small bug fix for admins.
If an admin launched tasks on behalf of another
normal user, the task form would only show the ToolConfigs
that the normal user has access. Which could lead to
an empty selection, and a crash in the interface.

The current change checks/assigns the TCs that can be
accessed by the current_user. This means a task could
potentially be launched on behalf of a user using a
ToolConfig that the user doesn't normally see, but
that's OK.