CleanCache BAC now created by portal
The old code used to send a message to each Bourreau with
a bunch of params to identify which files to clean. Then the
bourreau would create a BAC object.

In this new code, the portal no longer contacts the Bourreau,
it just creates the BAC directly. Lots of old code was cleaned.