New optional 'reverse ssh' service configuration. (#1532)
There is a new panel in the Bourreau config, and the boot system for them
has been adjusted. An admin can now explicitely specify where there database and SSH agent will come from. The Portal will maintain the ActiveResource connection only for the messages in there, and several portals can all talk to the same bourreau this way.