Adjusted task worker scheduler
Now if a subworker has already spent 30 minutes handling tasks,
it exits so that a re-evaluation of the active task list
is performed. This is to prevent a situation such that
a really large set of tasks are active and each take a very
long time to handle (e.g. setting up) and other tasks in
the meantime have finished processing and are waiting
to be reaped.