New way to specify max active tasks for users.
The old system generated a large ugly table in the Bourreau
show page, and stored the values in the meta data store.

This new way stores the values within the CpuQuota records.

A migration copies the values. Eventually we might want to
clean up the meta datastore too, but we'll introduce a cleaning
migration after this has been in production for some times.