Added stricter validations of Strings in BTQ forms
By default, BTQ inputs of type String must match a
rather strict list of characters.

An exception can be configured in the "custom" section
of the descriptor, on a case by case basis.

The core validation code already supports a bunch of
pre-defined cases to help out, e.g. ":basename:"