No longer restart TaskWorkers when launching tasks
It used to be that whenever a task was launched on
a Bourreau, a "start task" command was sent too. Which
meant that TaskWorkers were created if they were not
there already. This is not what we want, during maintenance
it is possible we want to keep TaskWorkers off even while
allowing users to submit tasks.

The new code just sends a "wake up" message instead,
which will wake up workers if they are there, but will
not create them if they don't exist.