Bindmount of plugins folder "container_mnt".
If a plugins contain a folder named "container_mnt",
and a boutiques descriptor has an entry in its custom
field called "cbrain:plugins-container-bindmount" with
a path as a value, then the plugins folder will be
bindmounted into the apptainer container under the
path specified in the descriptor.

Implements issue #1569 on GitHub.