Spurious params framework (#1581)
Adds a new controller directive, 'spurious_params_ban_ip'.

In a controller it would look like this:

spurious_params_ban_ip :credits => [], :about_us => []
spurious_params_ban_ip :credits  # same as with an explicit => []

spurious_params_ban_ip :login => [ :username, :password ]

spurious_params_ban_ip :new   => []

You provide to it the names of actions in the current controller, and a list of allowable parameters. If a GET or POST request is made with a bunch of unknown parameters, we know it's not a genuine client, so internally CBRAIN will ban the client's address. This entire mechanism is meant to be useful only for public-facing pages, not for for anything that happens to a logged-in client.