Added migration code from old to new BTQ
This commit contains a rake task to perform
a migration from the old Boutiques integrator
to the new one. At the end, for each tool migrated,
the JSON file for the descriptor is copied
to the 'boutiques_descriptor/' folder and it is
the responsibility of the admin to commit it
into the plugin's repo.